+++ src/types.ts
@@ -1,4 +1,14 @@
-import { ActionConfig } from "custom-card-helpers";
+// Attribution: to all devs at https://github.com/custom-cards/custom-card-helpers that let me use HomeAssistant object and its deps. Thanks alot!
+
+import {
+  HassEntities,
+  HassConfig,
+  Auth,
+  Connection,
+  MessageBase,
+  HassServices,
+  HassServiceTarget,
+} from "home-assistant-js-websocket";
 
 export interface CanvasGuageCardConfig {
   entity?: string;
@@ -11,8 +21,147 @@ export interface CanvasGuageCardConfig {
   card_height?: string;
   card_top?: string;
   card_left?: string;
+  dropshadow?: boolean;
 
   shadow_height?: string;
   shadow_bottom?: string;
   background_color?: string;
+}
+
+export interface ServiceCallRequest {
+  domain: string;
+  service: string;
+  serviceData?: Record<string, any>;
+  target?: HassServiceTarget;
+}
+export enum NumberFormat {
+  language = "language",
+  system = "system",
+  comma_decimal = "comma_decimal",
+  decimal_comma = "decimal_comma",
+  space_comma = "space_comma",
+  none = "none",
+}
+
+export enum TimeFormat {
+  language = "language",
+  system = "system",
+  am_pm = "12",
+  twenty_four = "24",
+}
+
+export interface FrontendLocaleData {
+  language: string;
+  number_format: NumberFormat;
+  time_format: TimeFormat;
+}
+
+export interface Theme {
+  // Incomplete
+  "primary-color": string;
+  "text-primary-color": string;
+  "accent-color": string;
+}
+
+export interface Themes {
+  default_theme: string;
+  themes: { [key: string]: Theme };
+}
+
+export interface Panel {
+  component_name: string;
+  config: { [key: string]: any } | null;
+  icon: string | null;
+  title: string | null;
+  url_path: string;
+}
+
+export interface Panels {
+  [name: string]: Panel;
+}
+
+export interface Resources {
+  [language: string]: { [key: string]: string };
+}
+
+export interface Translation {
+  nativeName: string;
+  isRTL: boolean;
+  fingerprints: { [fragment: string]: string };
+}
+
+export interface ServiceCallRequest {
+  domain: string;
+  service: string;
+  serviceData?: Record<string, any>;
+  target?: HassServiceTarget;
+}
+
+export type LocalizeFunc = (key: string, ...args: any[]) => string;
+
+export interface Credential {
+  auth_provider_type: string;
+  auth_provider_id: string;
+}
+
+export interface MFAModule {
+  id: string;
+  name: string;
+  enabled: boolean;
+}
+
+export interface CurrentUser {
+  id: string;
+  is_owner: boolean;
+  is_admin: boolean;
+  name: string;
+  credentials: Credential[];
+  mfa_modules: MFAModule[];
+}
+
+export interface HomeAssistant {
+  auth: Auth;
+  connection: Connection;
+  connected: boolean;
+  states: HassEntities;
+  services: HassServices;
+  config: HassConfig;
+  themes: Themes;
+  selectedTheme?: string | null;
+  panels: Panels;
+  panelUrl: string;
+
+  language: string;
+  locale: FrontendLocaleData;
+  // local stored language, keep that name for backward compability
+  selectedLanguage: string | null;
+  resources: Resources;
+  localize: LocalizeFunc;
+  translationMetadata: {
+    fragments: string[];
+    translations: {
+      [lang: string]: Translation;
+    };
+  };
+
+  dockedSidebar: boolean;
+  moreInfoEntityId: string;
+  user: CurrentUser;
+  callService: (
+    domain: ServiceCallRequest["domain"],
+    service: ServiceCallRequest["service"],
+    serviceData?: ServiceCallRequest["serviceData"],
+    target?: ServiceCallRequest["target"]
+  ) => Promise<void>;
+  callApi: <T>(
+    method: "GET" | "POST" | "PUT" | "DELETE",
+    path: string,
+    parameters?: { [key: string]: any }
+  ) => Promise<T>;
+  fetchWithAuth: (
+    path: string,
+    init?: { [key: string]: any }
+  ) => Promise<Response>;
+  sendWS: (msg: MessageBase) => Promise<void>;
+  callWS: <T>(msg: MessageBase) => Promise<T>;
 }