+++ RELEASENOTES.md
@@ -18,3 +18,8 @@ gauge:
 # 0.0.6
 
 Update to latest guage-card package to 2.1.7
+
+# 0.7.0
+- Dropshadow is now optional
+- Removed dependency on custom-card-helpers
+- Added devconainer that let you run HA and debug card